+++ packages/fixie/CHANGELOG.md
@@ -1,5 +1,13 @@
 # fixie
 
+## 7.0.11
+
+### Patch Changes
+
+- Fix DELETE operations to not expect content.
+- Updated dependencies
+  - @fixieai/fixie-common@1.0.11
+
 ## 7.0.10
 
 ### Patch Changes